backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]>=0.24.0
pydantic==2.5.0
python-multipart>=0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
tensorflow>=2.20.0
transformers>=4.35.0
torch>=2.1.0
spacy>=3.7.2
google-cloud-speech>=2.21.0
google-generativeai>=0.3.1,<0.4.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.3,<2.0.0
nltk>=3.8.1
shap>=0.43.0
lime>=0.2.0.1
requests>=2.31.0
aiofiles>=23.2.1
opencv-python>=4.8.0
librosa>=0.10.1
soundfile>=0.12.1
plotly>=5.17.0
seaborn>=0.12.2
matplotlib>=3.7.2
mangum>=0.17.0
# Additional ML/AI libraries for dynamic functionality
sentence-transformers>=2.2.2
langchain>=0.0.350
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
openai-whisper>=20231117
pyttsx3>=2.90
streamlit>=1.28.0
joblib>=1.3.2
pydub>=0.25.1
xgboost>=2.0.0
lightgbm>=4.1.0
selenium>=4.15.0
lxml>=4.9.3
beautifulsoup4>=4.12.2
PyPDF2>=3.0.0
python-docx>=0.8.11
